samsung-chagallwifi: select firmware pkg variant in device package (MR 4721)
We want both chagallwifi and chagalllte to be able to depend on
firmware-samsung-chagall, for both their downstream and mainline
kernel variants.

Previously we used install_if like:
install_if="device-samsung-chagallwifi-kernel-mainline"
which will only work for chagallwifi and not chagalllte.  Instead we
add two nonfree subpackages to the device packages that install_if's
depending on kernel choice.