linux-postmarketos-amlogic: add support for cottonwood boards (MR 4467)
This includes the "Alta" and "Solitude" boards, the latter of which I
also introduce a device package for in a later patch. It should be easy
to introduce support for the former, but I don't have the hardware to
test that on.

The two new patches have been merged and will be included in Linux 6.7
as I understand.