samsung-klimtlte: select firmware pkg variant in device package (MR 4721)
We want both klimtwifi and klimtlte to be able to depend on
firmware-samsung-klimt, for both their downstream and mainline kernel
variants.

Previously we used install_if like:
install_if="device-samsung-klimtlte-kernel-mainline" which will only
work for klimtlte and not klimtwifi.  Instead we add two nonfree
subpackages to the device packages that install_if's depending on
kernel choice.