samsung-chagallwifi: enable FIMC IS with downstream kernel (MR 4721)
As for several other exynos devices we need to disable the
video4linux's udev rule OR disable FIMC IS and patch kernel to fix
build error without FIMC IS.

Add dependency on postmarketos-base-downstream-disable-udev-v4l to
override the v4l rule with symlink to /dev/null.  We can then drop
FIMC IS patch and keep kernel config more similar to original one.