linux-lenovo-21bx: build specific commit, not branch archive (MR 4735)
I was building a branch archive previously, which meant that rebuilding
would fail if someone pushed to the branch upstream (the archive
checksum would change)... so I've reverted back to building from a
specific commit for now.

pkgrel was increased because this commit includes a couple of
audio-related patches that weren't included in the previous packaging of
the 6.7 branch.

[ci:skip-build] Already build-tested on Gitlab CI