linux-postmarketos-exynos7870: upgrade to 6.8.1 (MR 4977)
Upgrade the kernel to version 6.8.1.

Move the devicetrees to be built in-tree. This enables better
integration of the devicetree sources with the kernel, such as being
able to access macros available in include/dt-bindings.

Additionally, remove QCDT generation from the kernel. This is required
for supporting multiple devices, having different QCDTs. It can instead
be done in a per-device basis using boot-deploy.

[ci:skip-build]: already built successfully in CI