epson-embt2ws: new device (MR 5210)
Currently it requries the check_code result check in the
vendor uboot to be disabled as explained in the wiki.

Booting directly via reconfigured boot order still
does not work. It hangs somewhere in the kernel.

Firmware for Wifi/BT is not included yet.

It could use the linux-postmarketos-omap kernel in the future,
but there are still some dirty things in there e.g. for sound
which are not suitable for a multi-device kernel.
[ci:skip-build]: already built successfully in CI