kconfigcheck.toml: add Librem 5 category (MR 5562)
We've been disabling CONFIG_RS9116_FLASH_MODE for years at this point,
so this doesn't do anything but enforce the status quo.

Rationale for this can be found in the commit
b0b9e9718c26bb2d63d232ed0db5ec323a5eeb66.

This is deliberately not added to the community check as it is
device-specific.

Closes https://gitlab.com/postmarketOS/pmaports/-/issues/3157
[ci:skip-build]: already built successfully in CI