linux-postmarketos-exynos4: Build WM8994 MFD driver back into the kernel (MR 5618)
Viciouss noticed that audio no longer works after the recent changes
to build some things as modules, including the WM8994 driver.

Setting just the MFD bits to =y seems to fix it, so do it to fix audio.
Luckily these are pretty small and don't increase the kernel size much,
and we can still keep the actual sound driver (and thus all the sound
framework parts) compiled as modules.
[ci:skip-build]: already built successfully in CI