device-htc-m8: add lk2nd dependency (MR 5919)
Like most msm8974 devices, lk2nd is used to boot mainline linux on the
HTC One (M8).

Add it as a dependency of the mainline kernel build, so that it is built
when required. It can be flashed with 'pmbootstrap flasher flash_lk2nd'
while in the vendor bootloader fastboot mode.

Only extlinux mode is supported for mainline kernel, as otherwise the
initramfs is too big for the boot partition.

Signed-off-by: Alexandre Messier <alex@me.ssier.org>