main/postmarketos-initramfs: add missing USB gadget modules (MR 5795)
init_functions.sh makes use of the mass_storage and acm USB gadget
functions for error reporting, but the necessary kernel modules are not
included in the initramfs at the moment. Because of this, the functionality
works only on kernels where all these USB gadget functions are built-in and
not built as modules.

Add the missing modules and fix the sort order while at it.

[ci:skip-build]: bug in CI prevents build testing armv7, succeeded locally