mrhlpr.json: drop the unknown regex (MR 5840)
This regex seems to have served to be a bit more flexible than the
actual pass one, for any kind of exceptions needed. However, it
resulted in basically every MR since the introduction of the
COMMITSTYLE, and most likely before, to be caught by that regex
instead of the pass one. As we have made the pass regex more flexible
now, we can now remove this. This should help keep commit messages
more standardized. If there's a reason for a commit to be different,
then the merger can just ignore it.