main/postmarketos-ui-xfce4: set background for login greeter (MR 5889)
Actually Xfce4 is capable to use the "draw-user-backgrounds" option of
"slick-greeter", applying the users wallpaper on the greeter.

However, as it turns out, the users background in pmOS Xfce4 is set at first
login of the user. Therefore, the greeters background is black at the first
boot.

Therefore set the greeters background to the default postmarketOS wallpaper.